+++ config/text/about/fr/credits.yml
@@ -1,4 +1,4 @@
-label: Credits
+label: Crédits
 cards:
   - header: Développement
     body: |
