+++ config/text/about/fr/dve.yml
@@ -1,4 +1,4 @@
-label: Design Value Explorer
+label: l’Explorateur de valeurs de calcul
 cards:
   - header: Description
     body: |
