+++ config/text/about/fr/team.yml
@@ -1,4 +1,4 @@
-label: Team
+label: Équipe
 cards:
     - header: '[Charles Curry](https://pacificclimate.org/about-pcic/people/charles-curry)'
       body: Responsable de la science du climat.
